requirements.txt
Flask==3.0.0
tensorflow-cpu==2.15.0
Pillow==10.2.0
numpy==1.26.4
gunicorn==21.2.0